+++ src/ir/abstract.h
@@ -122,7 +122,6 @@ inline UnaryOp getUnary(Type type, Op op) {
     }
     case Type::v128:
     case Type::funcref:
-    case Type::externref:
     case Type::anyref:
     case Type::eqref:
     case Type::i31ref:
@@ -295,7 +294,6 @@ inline BinaryOp getBinary(Type type, Op op) {
     }
     case Type::v128:
     case Type::funcref:
-    case Type::externref:
     case Type::anyref:
     case Type::eqref:
     case Type::i31ref: